+++ src/accountant/AccountantExecutor.sol
@@ -61,7 +61,7 @@ contract AccountantExecutor is AccessControlUpgradeable {
     /// @notice Trigger an exchange rate update on the Accountant.
     /// @param newRate The new exchange rate to push
     /// @param computeTimestamp Off-chain computation timestamp for staleness check
-    function executeUpdateRate(uint256 newRate, uint256 computeTimestamp) external onlyRole(BOT_ROLE) {
+    function executeUpdateRate(uint64 newRate, uint64 computeTimestamp) external onlyRole(BOT_ROLE) {
         accountant.updateExchangeRate(newRate, computeTimestamp);
         emit RateUpdateExecuted(msg.sender, newRate, computeTimestamp);
     }
